+++ app/Http/Controllers/CustomerController.php
@@ -4,6 +4,7 @@ namespace App\Http\Controllers;
 
 use App\Models\Customer;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\DB;
 use Illuminate\Support\Facades\Validator;
 
 class CustomerController extends Controller
@@ -111,6 +112,15 @@ class CustomerController extends Controller
 
     public function destroy(string $id, Request $request)
     {
+        $customercount = DB::table('jual')->where('customer_id', $id)->count();
+
+        if ($customercount > 0) {
+            return response()->json([
+                'success' => false,
+                'message' => 'Tidak dapat dihapus, Data masih digunakan',
+            ]);
+        }
+
         customer::find($id)->delete();
 
         if ($request->ajax()) {
